báo cáo thực tập doanh nghiệp

+++ backend/controllers/foodController.js
@@ -38,6 +38,7 @@ const listFood = async (req, res) =>{
 const removeFood = async (req, res) =>{
     try {
         const food = await foodModel.findById(req.body.id)
+        if(!food) return res.json({success: false, message:"Food not found"})
         fs.unlink(`uploads/${food.image}`,() => {})
 
         await foodModel.findByIdAndDelete(req.body.id)
